# Optimised Ring.ShuffleShard() and disabled subring cache in store-gateway, ruler and compactor