… CHANGELOG.md
… pkg/compactor/compactor_ring_test.go
… pkg/ring/model_test.go
… pkg/ring/ring_test.go
… pkg/ring/util_test.go
… pkg/ruler/ruler.go
… pkg/ruler/ruler_replication_strategy.go
… pkg/ruler/ruler_ring.go
… pkg/storegateway/gateway.go
… pkg/storegateway/gateway_ring.go
… pkg/storegateway/replication_strategy.go
… pkg/storegateway/sharding_strategy.go
… pkg/storegateway/sharding_strategy_test.go